Move custom-snippet command to separate module
